song: fix warnings.

+++ game/song.cc
@@ -177,8 +177,8 @@ VocalTrack& Song::getVocalTrack(std::string vocalTrack) {
 	}
 }
 
-VocalTrack& Song::getVocalTrack(size_t idx) {
-	if (idx >= vocalTracks.size()) {
+VocalTrack& Song::getVocalTrack(unsigned idx) {
+	if (idx >= static_cast<unsigned>(vocalTracks.size())) {
 		throw std::logic_error("Index " + std::to_string(idx) + " out of bounds in Song::getVocalTrack (size: " + std::to_string(vocalTracks.size()) +").");
 		}
 	VocalTracks::iterator it = vocalTracks.begin();
@@ -191,7 +191,7 @@ double Song::getDurationSeconds() {
 		AVFormatContext *pFormatCtx = avformat_alloc_context();
 		if (avformat_open_input(&pFormatCtx, music["background"].string().c_str(), nullptr, nullptr) == 0) {
 			avformat_find_stream_info(pFormatCtx, nullptr);
-			m_duration = pFormatCtx->duration / AV_TIME_BASE;
+			m_duration = static_cast<double>(pFormatCtx->duration) / static_cast<double>(AV_TIME_BASE);
 			avformat_close_input(&pFormatCtx);
 			avformat_free_context(pFormatCtx);
 			return m_duration;
